// src/app/models/resource_models/IBaseResourceModel.ts
interface IBaseResourceModel {
  data?: any;
  message?: string;
  timestamp?: string;
  status?: number;
  isSuccess?: boolean;
}